supply_chain_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
neo4j==5.28.1
google-genai==1.0.0
numpy==1.26.4
pandas==2.2.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
orjson==3.10.7
